requirements.txt
pinecone-haystack
haystack-ai
fastapi
uvicorn
python-dotenv
pathlib
pypdf
sentence-transformers
pinecone

-e .